+++ libs/components/ag-grid/src/lib/modules/ag-grid/cell-renderers/cell-renderer-template/cell-renderer-template.component.spec.ts
@@ -1,4 +1,4 @@
-import { TemplateRef } from '@angular/core';
+import { TemplateRef, signal } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SkyAgGridCellRendererTemplateComponent } from './cell-renderer-template.component';
@@ -31,6 +31,26 @@ describe('SkyAgGridCellRendererTemplateComponent', () => {
             template: {} as TemplateRef<any>,
           },
         },
+      } as any);
+    }).not.toThrow();
+    expect(() => {
+      component.refresh({
+        data: { id: '1', name: 'test' },
+        value: 'test',
+        context: {},
+        colDef: {},
+      } as any);
+    }).not.toThrow();
+  });
+
+  it('should render template signal', () => {
+    const template = signal<TemplateRef<unknown> | undefined>(undefined);
+    expect(() => {
+      component.agInit({
+        data: { id: '1', name: 'test' },
+        value: 'test',
+        context: {},
+        colDef: { cellRendererParams: { template } },
       } as any);
     }).not.toThrow();
     expect(() => {