+++ libs/components/code-examples/src/lib/modules/ag-grid/data-entry-grid/data-manager-added/example.component.ts
@@ -2,9 +2,8 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
-  OnDestroy,
-  OnInit,
   inject,
+  signal,
 } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import {
@@ -19,7 +18,7 @@ import {
 import { SkyListSummaryModule } from '@skyux/lists';
 import { SkyModalConfigurationInterface, SkyModalService } from '@skyux/modals';
 
-import { Subject, map, takeUntil } from 'rxjs';
+import { map } from 'rxjs';
 
 import { AG_GRID_DEMO_DATA, AgGridDemoRow } from './data';
 import { EditModalContext } from './edit-modal-context';
@@ -45,40 +44,12 @@ const SOURCE_ID = 'data_entry_grid_data_manager_example_id';
     SkyListSummaryModule,
   ],
 })
-export class AgGridDataEntryGridDataManagerAddedExampleComponent
-  implements OnInit, OnDestroy
-{
-  protected items = AG_GRID_DEMO_DATA;
+export class AgGridDataEntryGridDataManagerAddedExampleComponent {
+  protected readonly items = signal<AgGridDemoRow[]>(AG_GRID_DEMO_DATA);
 
-  protected salesModal = SalesModalComponent;
+  protected readonly salesModal = SalesModalComponent;
 
-  #activeViewId = 'dataEntryGridWithDataManagerView';
-
-  #defaultDataState = new SkyDataManagerState({
-    filterData: {
-      filtersApplied: false,
-      filters: {},
-    },
-    views: [
-      {
-        viewId: 'dataEntryGridWithDataManagerView',
-        displayedColumnIds: [
-          'selected',
-          'context',
-          'name',
-          'age',
-          'startDate',
-          'endDate',
-          'department',
-          'jobTitle',
-          'validationCurrency',
-          'validationDate',
-        ],
-      },
-    ],
-  });
-
-  #dataManagerConfig = {
+  readonly #dataManagerConfig = {
     sortOptions: [
       {
         id: 'az',
@@ -95,8 +66,6 @@ export class AgGridDataEntryGridDataManagerAddedExampleComponent
     ],
   };
 
-  #ngUnsubscribe = new Subject<void>();
-
   readonly #changeDetectorRef = inject(ChangeDetectorRef);
   readonly #dataManagerSvc = inject(SkyDataManagerService);
   readonly #exampleSvc = inject(ExampleService);
@@ -110,36 +79,20 @@ export class AgGridDataEntryGridDataManagerAddedExampleComponent
   );
 
   constructor() {
-    this.#dataManagerSvc
-      .getActiveViewIdUpdates()
-      .pipe(takeUntil(this.#ngUnsubscribe))
-      .subscribe((activeViewId) => {
-        this.#activeViewId = activeViewId;
-        this.#changeDetectorRef.detectChanges();
-      });
-  }
-
-  public ngOnInit(): void {
     this.#dataManagerSvc.initDataManager({
-      activeViewId: this.#activeViewId,
+      activeViewId: 'dataEntryGridWithDataManagerView',
       dataManagerConfig: this.#dataManagerConfig,
-      defaultDataState: this.#defaultDataState,
+      defaultDataState: new SkyDataManagerState({}),
     });
-  }
-
-  public ngOnDestroy(): void {
-    this.#ngUnsubscribe.next();
-    this.#ngUnsubscribe.complete();
   }
 
   protected openModal(): void {
     const context = new EditModalContext();
-    context.gridData = this.items.slice();
+    context.gridData = structuredClone(this.items());
 
     this.#changeDetectorRef.markForCheck();
 
     const options: SkyModalConfigurationInterface = {
-      ariaDescribedBy: 'docs-edit-grid-modal-content',
       providers: [
         {
           provide: EditModalContext,
@@ -155,10 +108,8 @@ export class AgGridDataEntryGridDataManagerAddedExampleComponent
       if (result.reason === 'cancel' || result.reason === 'close') {
         alert('Edits canceled!');
       } else {
-        this.items = result.data as AgGridDemoRow[];
-        alert('Saving data!');
+        this.items.set(result.data as AgGridDemoRow[]);
       }
-
       this.#changeDetectorRef.markForCheck();
     });
   }