+++ libs/components/code-examples/src/lib/modules/ag-grid/data-grid/template-ref-column/example.component.ts
@@ -1,18 +1,8 @@
-import {
-  Component,
-  OnInit,
-  TemplateRef,
-  ViewChild,
-  inject,
-} from '@angular/core';
+import { Component, TemplateRef, inject, viewChild } from '@angular/core';
 import { SkyAgGridModule, SkyAgGridService, SkyCellType } from '@skyux/ag-grid';
 
 import { AgGridModule } from 'ag-grid-angular';
-import {
-  AllCommunityModule,
-  GridOptions,
-  ModuleRegistry,
-} from 'ag-grid-community';
+import { AllCommunityModule, ModuleRegistry } from 'ag-grid-community';
 
 import { AG_GRID_DEMO_DATA } from './data';
 
@@ -26,54 +16,45 @@ ModuleRegistry.registerModules([AllCommunityModule]);
   templateUrl: './example.component.html',
   imports: [AgGridModule, SkyAgGridModule],
 })
-export class AgGridDataGridTemplateRefColumnExampleComponent implements OnInit {
-  @ViewChild('boldColumn', { static: true })
-  protected boldColumn: TemplateRef<unknown> | undefined;
-
-  @ViewChild('emphasizedColumn', { static: true })
-  protected emphasizedColumn: TemplateRef<unknown> | undefined;
-
-  protected gridData = AG_GRID_DEMO_DATA;
-  protected gridOptions: GridOptions | undefined;
-
-  readonly #agGridSvc = inject(SkyAgGridService);
-
-  public ngOnInit(): void {
-    this.gridOptions = this.#agGridSvc.getGridOptions({
-      gridOptions: {
-        columnDefs: [
-          {
-            field: 'name',
-            headerName: 'Name',
-            initialWidth: 150,
+export class AgGridDataGridTemplateRefColumnExampleComponent {
+  protected readonly boldColumn = viewChild<TemplateRef<unknown>>('boldColumn');
+  protected readonly emphasizedColumn =
+    viewChild<TemplateRef<unknown>>('emphasizedColumn');
+  protected gridOptions = inject(SkyAgGridService).getGridOptions({
+    gridOptions: {
+      columnDefs: [
+        {
+          field: 'name',
+          headerName: 'Name',
+          initialWidth: 150,
+        },
+        {
+          field: 'age',
+          headerName: 'Age',
+          type: SkyCellType.Number,
+          maxWidth: 80,
+          resizable: false,
+        },
+        {
+          field: 'department',
+          headerName: 'Department',
+          type: SkyCellType.Template,
+          cellRendererParams: {
+            template: this.boldColumn,
           },
-          {
-            field: 'age',
-            headerName: 'Age',
-            type: SkyCellType.Number,
-            maxWidth: 80,
-            resizable: false,
+          initialWidth: 220,
+        },
+        {
+          field: 'jobTitle',
+          headerName: 'Title',
+          type: SkyCellType.Template,
+          cellRendererParams: {
+            template: this.emphasizedColumn,
           },
-          {
-            field: 'department',
-            headerName: 'Department',
-            type: SkyCellType.Template,
-            cellRendererParams: {
-              template: this.boldColumn,
-            },
-            initialWidth: 220,
-          },
-          {
-            field: 'jobTitle',
-            headerName: 'Title',
-            type: SkyCellType.Template,
-            cellRendererParams: {
-              template: this.emphasizedColumn,
-            },
-            initialWidth: 220,
-          },
-        ],
-      },
-    });
-  }
+          initialWidth: 220,
+        },
+      ],
+      rowData: AG_GRID_DEMO_DATA,
+    },
+  });
 }