+++ CHANGELOG.md
@@ -1,6 +1,35 @@
 # Changelog
 
 
+## [13.11.3](https://github.com/blackbaud/skyux/compare/13.11.2...13.11.3) (2025-12-16)
+
+
+### Bug Fixes
+
+* **components/ag-grid:** wait for grid to be ready before api call ([#4129](https://github.com/blackbaud/skyux/issues/4129)) ([12e1f17](https://github.com/blackbaud/skyux/commit/12e1f17494a0654fc1ff6efc20dc377764fd4759)), closes [AB#3652352](https://github.com/AB/issues/3652352)
+
+
+
+## [13.11.2](https://github.com/blackbaud/skyux/compare/13.11.1...13.11.2) (2025-12-16)
+
+
+### Bug Fixes
+
+* **components/ag-grid:** wait for grid to be ready before applying data manager column sizes ([#4128](https://github.com/blackbaud/skyux/issues/4128)) ([4e28b1a](https://github.com/blackbaud/skyux/commit/4e28b1ad66acbd6485c3326f7616e3b57f475425)), closes [AB#3652352](https://github.com/AB/issues/3652352)
+* **components/data-manager:** cleanup observable subscription ([#4126](https://github.com/blackbaud/skyux/issues/4126)) ([b61a72b](https://github.com/blackbaud/skyux/commit/b61a72b2dbd54a373622c0d8635535ecca5f8227)), closes [AB#3597884](https://github.com/AB/issues/3597884)
+
+
+
+## [13.11.1](https://github.com/blackbaud/skyux/compare/13.11.0...13.11.1) (2025-12-15)
+
+
+### Bug Fixes
+
+* **components/code-examples:** simplify data manager code examples ([#4122](https://github.com/blackbaud/skyux/issues/4122)) ([77198c8](https://github.com/blackbaud/skyux/commit/77198c8dd31e1033b3e852da13031c6e6b495bae))
+* **components/docs-tools:** allow for code examples without specifying `componentType` ([#4124](https://github.com/blackbaud/skyux/issues/4124)) ([780643f](https://github.com/blackbaud/skyux/commit/780643fd837324c06753c72e2d898f3ef38ea776)), closes [AB#3583748](https://github.com/AB/issues/3583748)
+
+
+
 # [13.11.0](https://github.com/blackbaud/skyux/compare/13.10.0...13.11.0) (2025-12-10)